Style: Alpha,Background,Point를 랜덤 생성해주는 메서드를 따로 분리해 AlphaFactory, BackgroundColorFactory, PointFactory 객체로 분리
backgroundButton 클릭시 배경색을 랜덤 변환하는 코드를 재사용하기 위해 분리
